+++ CHANGELOG.md
@@ -23,9 +23,11 @@ For instructions on upgrading to newer versions, visit
 
 ### New Features
 
-* \#3846 Allow #pluck when none is used in criteria. (Braulio Martinez)
-
 * \#3985 Return nil when using `{upsert: true}` in `find_and_modify` (Adrien Siami)
+
+* \#3963 Allow extended JSON object ids to be passed to `find`.
+
+* \#3846 Allow #pluck when none is used in criteria. (Braulio Martinez)
 
 ### Resolved Issues